created usersController file and moved route logic to it
